@@ -2,23 +2,23 @@
   "name": "hammertime",
   "version": "1.0.0",
   "scripts": {
-    "predev:translation-credits": "ts-node ./utils/translation-credits.ts",
+    "predev:translation-credits": "tsx ./utils/translation-credits.ts",
     "predev": "concurrently -p name npm:predev:*",
     "dev": "next",
     "build": "next build",
-    "build:previews": "ts-node utils/build-social-previews.ts",
+    "build:previews": "tsx utils/build-social-previews.ts",
     "start": "next start",
     "lint": "next lint",
     "format": "prettier --write .",
     "prepare": "husky install"
   },
   "dependencies": {
-    "@fortawesome/fontawesome-common-types": "^6.1.0",
-    "@fortawesome/fontawesome-svg-core": "^6.1.0",
-    "@fortawesome/free-brands-svg-icons": "^6.1.0",
-    "@fortawesome/free-regular-svg-icons": "^6.1.0",
-    "@fortawesome/free-solid-svg-icons": "^6.1.0",
-    "@fortawesome/react-fontawesome": "^0.2.2",
+    "@fortawesome/fontawesome-common-types": "^7.2.0",
+    "@fortawesome/fontawesome-svg-core": "^7.2.0",
+    "@fortawesome/free-brands-svg-icons": "^7.2.0",
+    "@fortawesome/free-regular-svg-icons": "^7.2.0",
+    "@fortawesome/free-solid-svg-icons": "^7.2.0",
+    "@fortawesome/react-fontawesome": "^3.2.0",
     "@kfonts/bm-euljiro": "^0.1.3",
     "@mantine/core": "^6.0.22",
     "@mantine/dates": "^6.0.22",
@@ -30,19 +30,17 @@
     "classnames": "^2.2.6",
     "cookies-next": "^6.1.1",
     "dayjs": "^1.11.13",
-    "i18next": "^23.15.1",
     "lodash": "^4.17.21",
     "moment": "^2.30.1",
     "moment-timezone": "^0.6.0",
-    "next": "^15.5.7",
-    "next-i18next": "^15.4.3",
-    "next-seo": "^6.6.0",
+    "next": "^16.1.6",
+    "next-intl": "^4.8.3",
+    "next-seo": "^6.8.0",
     "noto-sans-japanese": "^1.0.0",
     "postcss-import": "^13.0.0",
     "react": "^18",
     "react-dom": "^18",
     "react-flagkit": "^2.0.4",
-    "react-i18next": "^15.0.2",
     "sass": "^1.79.4",
     "stylis": "^4.3.4",
     "stylis-plugin-rtl": "^2.1.1",
@@ -77,16 +75,18 @@
     "imagemin": "^9.0.0",
     "imagemin-pngquant": "^10.0.0",
     "lint-staged": "^10.5.3",
-    "prettier": "^2.2.1",
+    "prettier": "^3.8.1",
     "puppeteer": "^23.4.1",
+    "sass": "~1.70.0",
     "sharp": "^0.33.5",
     "source-code-pro": "^2.38.0",
-    "ts-node": "^10.9.1",
     "tsimportlib": "^0.0.5",
+    "tsx": "^4.21.0",
     "typescript": "^5.1"
   },
   "license": "MIT",
   "lint-staged": {
     "**/*": "prettier --write --ignore-unknown"
-  }
+  },
+  "type": "module"
 }
