@@ -2,15 +2,18 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Alert, LoadingOverlay, useMantineTheme } from '@mantine/core';
 import { LoadingIndicator } from 'components/app/LoadingIndicator';
 import { useLocalSettings } from 'components/contexts/LocalSettingsProvider';
-import { useTranslation } from 'next-i18next';
-import { CSSProperties, FC } from 'react';
+import { useTranslations } from 'next-intl';
 import { useRouter } from 'next/router';
+import { CSSProperties, FC } from 'react';
 
 const centerAlignedStyle: CSSProperties = {
   display: 'flex',
   alignItems: 'center',
   justifyContent: 'center',
   marginBottom: '1rem',
+  minHeight: '100vh',
+  minWidth: '100vw',
+  zIndex: 10,
 };
 
 const transitionDuration = 500;
@@ -18,7 +21,7 @@ const transitionDuration = 500;
 export const AppLoadingOverlay: FC = () => {
   const theme = useMantineTheme();
   const router = useRouter();
-  const { t } = useTranslation();
+  const t = useTranslations();
   const { sidebarOnRight, sidebarOffDesktop, customInputEnabled, combinedInputsEnabled } = useLocalSettings();
 
   const visible = [sidebarOnRight, sidebarOffDesktop, customInputEnabled, combinedInputsEnabled, router.isReady ? true : null].some(
@@ -39,8 +42,8 @@ export const AppLoadingOverlay: FC = () => {
           </div>
 
           <noscript>
-            <Alert title={t('common:jsDisabled.title')} icon={<FontAwesomeIcon icon="exclamation-triangle" fixedWidth />} color="red">
-              {t('common:jsDisabled.body')}
+            <Alert title={t('jsDisabled.title')} icon={<FontAwesomeIcon icon="exclamation-triangle" fixedWidth />} color="red">
+              {t('jsDisabled.body')}
             </Alert>
           </noscript>
         </>
