@@ -2,7 +2,8 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { Alert } from '@mantine/core';
 import { getCookie, setCookie } from 'cookies-next';
 import moment from 'moment-timezone';
-import { useTranslation } from 'next-i18next';
+import { useTranslations } from 'next-intl';
+import { useRouter } from 'next/router';
 import { FC, useEffect, useMemo, useState } from 'react';
 import { useLocale } from 'src/util/common';
 
@@ -10,10 +11,8 @@ const howToCookieName = 'how-to-dismiss';
 const howToCookieValue = 'how-to-dismiss';
 
 export const HowToAlert: FC = () => {
-  const {
-    t,
-    i18n: { language },
-  } = useTranslation();
+  const t = useTranslations();
+  const { locale: language } = useRouter();
   const locale = useLocale(language);
   const [showHowTo, setShowHowTo] = useState(false);
   const handleHowToClose = () => {
@@ -24,7 +23,7 @@ export const HowToAlert: FC = () => {
   };
 
   const syntaxColName = useMemo(() => {
-    const originalText = t('common:table.syntax');
+    const originalText = t('table.syntax');
     // Lowercase column name in text only for this language
     if (locale === 'pt-br') {
       return originalText.toLowerCase();
@@ -40,13 +39,13 @@ export const HowToAlert: FC = () => {
 
   return (
     <Alert
-      title={t('common:seoDescription')}
+      title={t('seoDescription')}
       icon={<FontAwesomeIcon icon="info" fixedWidth />}
       color="dark"
       withCloseButton
       onClose={handleHowToClose}
     >
-      {t('common:howTo', { syntaxColName })}
+      {t('howTo', { syntaxColName })}
     </Alert>
   );
 };
