@@ -1,7 +1,7 @@
 import { faCheck, faExclamationTriangle } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { notifications } from '@mantine/notifications';
-import { TFunction } from 'i18next';
+import { TFunction } from 'src/config';
 import moment from 'moment-timezone';
 import { useEffect } from 'react';
 
@@ -14,7 +14,7 @@ export const useServerTimeSync = (t: TFunction) => {
     let mounted = true;
     notifications.show({
       id: notificationId,
-      message: t('common:timeSync.syncing'),
+      message: t('timeSync.syncing'),
       withCloseButton: false,
       loading: true,
     });
@@ -28,7 +28,7 @@ export const useServerTimeSync = (t: TFunction) => {
       if (Math.abs(deltaTimeMs) > fiveMinutesInMs) {
         notifications.update({
           id: notificationId,
-          message: t('common:timeSync.outOfSync', { deltaTimeMs }),
+          message: t('timeSync.outOfSync', { deltaTimeMs }),
           withCloseButton: true,
           icon: <FontAwesomeIcon icon={faExclamationTriangle} />,
           color: 'yellow',
@@ -36,7 +36,7 @@ export const useServerTimeSync = (t: TFunction) => {
       } else {
         notifications.update({
           id: notificationId,
-          message: t('common:timeSync.inSync', { deltaTimeMs }),
+          message: t('timeSync.inSync', { deltaTimeMs }),
           withCloseButton: true,
           autoClose: 2e3,
           icon: <FontAwesomeIcon icon={faCheck} />,
