@@ -1,6 +1,9 @@
-import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
-import type { AppI18nNamespaces } from 'react-i18next';
-import { DEFAULT_I18N_NAMESPACES } from 'src/config';
+import { ServerSideTranslations } from 'src/config';
 
-export const typedServerSideTranslations = (locale?: string, keys?: AppI18nNamespaces[]) =>
-  serverSideTranslations(locale as string, keys ? [...DEFAULT_I18N_NAMESPACES, ...keys] : DEFAULT_I18N_NAMESPACES);
+export const typedServerSideTranslations = async (locale?: string): Promise<ServerSideTranslations> => {
+  const languageCode = locale ?? 'en-GB';
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment
+  const messages = await import(`src/../public/locales/${languageCode}/common.json`);
+  // eslint-disable-next-line @typescript-eslint/no-unsafe-assignment,@typescript-eslint/no-unsafe-member-access
+  return { messages: { ...messages.default } };
+};
